--- src/__mocks__/add-delete.js
@@ -1,9 +1,0 @@
-import localStorageMock from "./localStorage.js";
-
-function addTodoToList(description, completed = false, id, list) {
-  const todo = { description, completed, id };
-  list.push(todo);
-  return list;
-}
-
-module.exports = { addTodoToList };
