+++ src/add.js
@@ -1,0 +1,5 @@
+export function addTodoToList(description, completed = false, id, list) {
+  const todo = { description, completed, id };
+  list.push(todo);
+  return list;
+}
