+++ src/delete.js
@@ -1,12 +1,9 @@
-import localStorageMock from "../__mocks__/localstorage";
-
 export function deleteTask(idx) {
-  const storage = JSON.parse(localStorageMock.getItem("todos"));
+  const storage = JSON.parse(localStorage.getItem('todos'));
   const savedata = storage.filter((value, index) => index !== idx);
   savedata.forEach((obj, index) => {
     obj.index = index + 1;
   });
-  localStorageMock.setItem("todos", JSON.stringify(savedata));
+  localStorage.setItem('todos', JSON.stringify(savedata));
   window.location.reload();
 }
-
