+++ src/test/delete.test.js
@@ -1,6 +1,7 @@
 import { deleteTask } from '../__mocks__/delete';
+import localStorage from '../__mocks__/localstorage';
+
 const jsdom = require('jsdom');
-import localStorage from "../__mocks__/localstorage";
 
 const { JSDOM } = jsdom;
 
@@ -9,23 +10,23 @@ const document = dom.window.document; // eslint-disable-line
 const window = dom.window.document;
 // Object.defineProperty(window, 'localStorage', { value: localStorageMock });
 
-describe ('removing list from local storage', ()=>{
-    const list = [
-      {
-        index: 0,
-        description: 'look outside',
-        completed: true
-      },
-      {
-        index: 1,
-        description: 'Add homework',
-        completed: true,
-      },
-    ];
-    localStorageMock.setItem('todos', JSON.stringify(list));
-    test ('test the description',()=> {
-      deleteTask(1);
-      const storage = JSON.parse(localStorageMock.getItem('todos'))
-      expect(storage).toHaveLength(1);
-    })
-  })
+describe('removing list from local storage', () => {
+  const list = [
+    {
+      index: 0,
+      description: 'look outside',
+      completed: true,
+    },
+    {
+      index: 1,
+      description: 'Add homework',
+      completed: true,
+    },
+  ];
+  localStorageMock.setItem('todos', JSON.stringify(list));
+  test('test the description', () => {
+    deleteTask(1);
+    const storage = JSON.parse(localStorageMock.getItem('todos'));
+    expect(storage).toHaveLength(1);
+  });
+});
